Recommender Systems Project
In this notebook, I focused on providing a basic recommendation system by suggesting items that are most similar to a particular item, in this case, movies. It is not a true robust recommendation system, to describe it more accurately,it just tells you what movies/items are most similar to your movie choice.