Keras API Project
Given historical data on loans given out with information on whether or not the borrower defaulted (charge-off), I build a model that can predict wether or nor a borrower will pay back their loan. This way in the future when I get a new potential customer I can assess whether or not they are likely to pay back the loan.